pyproject.toml
[project]
name = "vibe-spec"
version = "0.1.0"
description = "Spec-Driven Vibe Coding Framework - A hierarchical specification system for LLM-driven development"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Vibe Team"}
]
keywords = ["specifications", "llm", "vibe-coding", "testing", "contracts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "hypothesis>=6.0",
]

[project.scripts]
vibe-spec = "vibe_spec.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibe_spec"]

[tool.pytest.ini_options]
testpaths = ["tests", "examples/tests"]
